feat: add secrets
